Prep for userland processes

+++ avs-libc-static/include/_types/_time_t.h
@@ -1,7 +1,7 @@
 #ifndef __TIME_T_TYPE_DEF_H_
 #define __TIME_T_TYPE_DEF_H_
 
-typedef long long time_t;
+typedef uint32_t time_t;
 typedef long long suseconds_t;
 
 #define __suseconds_t_defined 1
