+++ avs-libc-static/include/_types/_wchar_t.h
@@ -2,7 +2,7 @@
 #define WCHAR_T_H_
 
 #ifndef __cplusplus
-typedef long wchar_t;
+typedef int wchar_t;
 #endif
 
 #endif // WCHAR_T_H_
